fix: Контекст хук и провайдер хранятся отдельно
