--- src/shared/lib/theme/ThemeContext.ts
@@ -1,26 +1,0 @@
-import { createContext, useContext, FC, ReactNode } from 'react';
-
-export type Theme = 'light' | 'dark';
-
-interface ThemeContextType {
-    theme: Theme;
-    toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-interface ThemeProviderProps {
-    children: ReactNode;
-}
-
-export const ThemeProvider: FC<ThemeProviderProps> = ({ children }) => {
-    return null;
-};
-
-export const useTheme = () => {
-    const context = useContext(ThemeContext);
-    if (context === undefined) {
-        throw new Error('err');
-    }
-    return context;
-};
