+++ src/shared/lib/theme/index.ts
@@ -1,3 +1,3 @@
-export { ThemeProvider } from './ThemeProvider';
+export { ThemeProvider } from './ThemeContext';
 export { useTheme } from './ThemeContext';
 export type { Theme } from './ThemeContext';
